+++ NEWS.md
@@ -2,6 +2,7 @@
  - March 7, 2023
  - fix `topTable()` to deal with multiple `coef` as array
  - add vignette about non-linear effects
+ - Add `colData<-` for `dreamletProcessedData`
 
 # dreamlet 0.0.63
  - March 3, 2023